+++ app/src/main/java/jp/ac/ritsumei/ise/phy/exp2is0667fe/danceevaluation/ApiService.java
@@ -5,5 +5,5 @@ import retrofit2.http.Body;
 import retrofit2.http.POST;
 public interface ApiService {
     @POST("results/sort")
-    Call<List<History_ApiResponseItem>> sortResults(@Body History_RequestBody body);
+    Call<List<HistoryApiResponseItem>> sortResults(@Body HistoryRequestBody body);
 }
